Fixing include for termcap
A problem happens when trying to use the GNU termcap library that is
distributed with bash.  The termcap.h file is not in the include path.
There doesn't seem to be an elegant way to inclde it in the Makefiles
without accruing more changes than I'd like, so this patch updates
readline to include the machines.h file and that file indicates if we
should use the GNU termcap or not.